+++ src/main/java/org/isf/menu/gui/MainMenu.java
@@ -57,6 +57,7 @@ import org.isf.menu.manager.UserBrowsingManager;
 import org.isf.menu.model.User;
 import org.isf.menu.model.UserGroup;
 import org.isf.menu.model.UserMenuItem;
+import org.isf.session.UserSession;
 import org.isf.sms.service.SmsSender;
 import org.isf.utils.exception.OHServiceException;
 import org.isf.utils.exception.gui.OHServiceExceptionUtil;
