+++ src/main/java/org/isf/menu/gui/SplashWindow3.java
@@ -34,6 +34,9 @@ import javax.swing.JWindow;
 import javax.swing.SwingUtilities;
 
 import org.isf.generaldata.GeneralData;
+import org.isf.session.LogoutEventListener;
+import org.isf.session.SessionRefreshedRunnable;
+import org.isf.session.UserSession;
 import org.isf.utils.jobjects.DelayTimer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
